Added astype('int64')

+++ pandapower/converter/powerfactory/validate.py
@@ -478,7 +478,7 @@ def _validate_pf_conversion_balanced(net, in_both, all_diffs):
         if bus_dc_diff.isna().all():
             print("Warning: bus_dc_diff contains only NaN values. Continuing without evaluating dc busses.")
         else:
-            bus_dc_id = abs(bus_dc_diff).abs().idxmax()
+            bus_dc_id = abs(bus_dc_diff).abs().idxmax().astype('int64')
             bus_dc_id = int(bus_dc_id)  # Konvertiere den Index in einen int
             logger.info("Maximum bus_dc vm_pu difference between pandapower and powerfactory: %.6f "
                         "p.u. at bus_dc %d (%s)" % (max(abs(bus_dc_diff)), bus_dc_id, net.bus_dc.at[bus_dc_id, 'name']))
